64-bit changes: casts, and ioctl cmds are u_longs
SCCS-vsn: 8.5

+++ sys/net/if.c
@@ -30,7 +30,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)if.c	8.4 (Berkeley) 07/01/94
+ *	@(#)if.c	8.5 (Berkeley) 01/09/95
  */
 
 #include <sys/param.h>
@@ -458,7 +458,7 @@ ifunit(name)
 int
 ifioctl(so, cmd, data, p)
 	struct socket *so;
-	int cmd;
+	u_long cmd;
 	caddr_t data;
 	struct proc *p;
 {
